chore: 페이지 첫 랜더링시 선수 상세데이터 안나오는 상황 해결 (타입 설정 미스)
