+++ src/app/player/pitcher/page.tsx
@@ -8,7 +8,10 @@ import Banner from '@/components/common/banner/banner'
 import { PLAYER_BANNER_DATA } from '@/contants/player'
 import TabMenu from '@/components/common/tab-menu/tab-menu'
 import { PlayerCode } from '@/types/player'
-import { getPitcherPlayerList } from '@/app/api/player/api'
+import {
+  getPitcherPlayerList,
+  getPitcherPlayerDetail,
+} from '@/app/api/player/api'
 
 interface PlayerCard {
   pcode: PlayerCode
@@ -17,8 +20,11 @@ interface PlayerCard {
 }
 
 export default function Pitcher() {
-  const [playerPcode, setPlayerPcode] = useState<PlayerCode>({ pcode: 53006 })
+  const [playerPcode, setPlayerPcode] = useState<PlayerCode>(53006)
   const [cards, setCards] = useState<PlayerCard[]>([])
+  const [detailData, setDetailData] = useState()
+  const [seasonData, setSeasonData] = useState()
+  const [playerImg, setPlayerImg] = useState()
 
   useEffect(() => {
     const fetchPitcherPlayerList = async () => {
@@ -32,6 +38,21 @@ export default function Pitcher() {
     fetchPitcherPlayerList()
   }, [])
 
+  useEffect(() => {
+    const fetchPitcherPlayerDetail = async () => {
+      try {
+        const data = await getPitcherPlayerDetail(playerPcode)
+        console.log('투수상세 API 요청 성공 : ', data)
+        setDetailData(data.data.gameplayer)
+        setPlayerImg(data.data.gameplayer.playerPrvwImg1)
+        setSeasonData(data.data.yearrecordlist[0])
+      } catch (error) {
+        console.error('getPlayerDetail 요청 실패 : ', error)
+      }
+    }
+    fetchPitcherPlayerDetail()
+  }, [playerPcode])
+
   return (
     <>
       <BannerTest />
@@ -42,7 +63,11 @@ export default function Pitcher() {
           </div>
 
           <div className="w-full flex-grow rounded-lg p-4 shadow-md md:w-4/5">
-            <PlayerDetail pcode={playerPcode} />
+            <PlayerDetail
+              detailData={detailData}
+              seasonData={seasonData}
+              playerImg={playerImg}
+            />
           </div>
         </div>
 
@@ -61,7 +86,6 @@ const BannerTest = () => {
         title="투수"
         subtitle="투수 관련 정보 및 데이터를 확인하세요"
       />
-
       <TabMenu tabs={PLAYER_BANNER_DATA['/player'].tabs} />
     </Banner>
   )
