+++ src/components/player/player-detail.tsx
@@ -3,32 +3,19 @@ import PlayerDetailData from './player-detail/player-detail-data'
 import PlayerImg from './player-detail/player-img'
 import PlayerSeasonData from './player-detail/player-season-data'
 import { useEffect, useState } from 'react'
-import { PlayerCode } from '@/types/player'
+import { PlayerData, PlayerSeasonDataProps } from '@/types/player'
 
 interface PlayerDetailProps {
-  pcode: PlayerCode
+  detailData: PlayerData
+  seasonData: PlayerSeasonDataProps
+  playerImg: string | undefined
 }
 
-export default function PlayerDetail({ pcode }: PlayerDetailProps) {
-  const [detailData, setDetailData] = useState()
-  const [seasonData, setSeasonData] = useState()
-  const [playerImg, setPlayerImg] = useState()
-
-  useEffect(() => {
-    const fetchPitcherPlayerDetail = async () => {
-      try {
-        const data = await getPitcherPlayerDetail(pcode)
-        console.log('API 요청 성공 : ', data)
-        setDetailData(data.data.gameplayer)
-        setPlayerImg(data.data.gameplayer.playerPrvwImg1)
-        setSeasonData(data.data.yearrecordlist[0])
-      } catch (error) {
-        console.error('getPlayerDetail 요청 실패 : ', error)
-      }
-    }
-    fetchPitcherPlayerDetail()
-  }, [pcode])
-
+export default function PlayerDetail({
+  detailData,
+  seasonData,
+  playerImg,
+}: PlayerDetailProps) {
   return (
     <div className="mx-auto flex max-w-6xl flex-col rounded-lg p-6 md:flex-row md:gap-12">
       <div className="flex w-full flex-col items-center space-y-4 md:w-1/2">
